back-end/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.2.1
bcrypt==5.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
